+++ core-ng/src/main/java/core/framework/internal/stat/Stats.java
@@ -1,7 +1,9 @@
 package core.framework.internal.stat;
 
+import core.framework.util.ASCII;
 import core.framework.util.Maps;
 
+import java.text.NumberFormat;
 import java.util.Map;
 
 /**
@@ -17,14 +19,24 @@ public class Stats {
         return "WARN";
     }
 
-    public void warn(String errorCode, String errorMessage) {
-        if (this.errorCode == null) {
-            this.errorCode = errorCode;
-            this.errorMessage = errorMessage;
+    public void put(String key, double value) {
+        stats.put(key, value);
+    }
+
+    public void checkHighUsage(double used, double max, double threshold, String name) {
+        double usage = used / max;
+        if (usage >= threshold) {
+            NumberFormat format = NumberFormat.getPercentInstance();
+            errorCode = "HIGH_" + ASCII.toUpperCase(name) + "_USAGE";
+            errorMessage = name + " usage is too high, usage=" + format.format(usage);
         }
     }
 
-    public void put(String key, double value) {
-        stats.put(key, value);
+    public void checkHighUsage(double usagePercent, double threshold, String name) {
+        if (usagePercent >= threshold) {
+            NumberFormat format = NumberFormat.getPercentInstance();
+            errorCode = "HIGH_" + ASCII.toUpperCase(name) + "_USAGE";
+            errorMessage = name + " usage is too high, usage=" + format.format(usagePercent);
+        }
     }
 }