+++ CHANGELOG.md
@@ -1,5 +1,5 @@
 ## Change log
-### 7.3.5 (1/30/2020 - )
+### 7.3.5 (1/30/2020 - 2/6/2020)
 * httpClient: support retry on 429 too many requests
 * monitor: added es monitoring, refer to https://github.com/neowu/core-ng-project/wiki/Ext for example config
 * log-processor: added realted es visualization and dashboard